backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.2.3
numpy==2.3.2
torch==2.8.0
torchvision==0.23.0
pillow==11.0.0
requests==2.32.3
